# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for managing EZFW (Easy Firewall/Proxy) configurations for Xray and Sing-box.",
        "title": "EZFW API",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/tools4net/ezfw/issues",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/configs/singbox": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves a paginated list of SingBox configurations.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ConfigsSingBox"
                ],
                "summary": "List SingBox Configs",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "example": 10,
                        "description": "Number of items to return per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "example": 0,
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of SingBox configurations",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/github_com_tools4net_ezfw_backend_internal_models.SingBoxConfig"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Failed to list configurations",
                        "schema": {
                            "$ref": "#/definitions/github_com_tools4net_ezfw_backend_internal_models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Adds a new SingBox configuration to the store.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ConfigsSingBox"
                ],
                "summary": "Create SingBox Config",
                "parameters": [
                    {
                        "description": "SingBox Configuration Object",
                        "name": "singBoxConfig",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tools4net_ezfw_backend_internal_models.SingBoxConfig"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created SingBox configuration",
                        "schema": {
                            "$ref": "#/definitions/github_com_tools4net_ezfw_backend_internal_models.SingBoxConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid input or missing name",
                        "schema": {
                            "$ref": "#/definitions/github_com_tools4net_ezfw_backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Failed to create configuration",
                        "schema": {
                            "$ref": "#/definitions/github_com_tools4net_ezfw_backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/configs/singbox/{configId}/generate": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves a SingBox configuration and returns it in a format suitable for the sing-box binary, stripping internal metadata.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ConfigsSingBox"
                ],
                "summary": "Generate SingBox Config File",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SingBox Configuration ID",
                        "name": "configId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The core SingBox configuration JSON. The actual structure depends on the config (models.SingBoxConfig without metadata).",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid or missing configuration ID",
                        "schema": {
                            "$ref": "#/definitions/github_com_tools4net_ezfw_backend_internal_models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found - Configuration not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_tools4net_ezfw_backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Failed to retrieve configuration",
                        "schema": {
                            "$ref": "#/definitions/github_com_tools4net_ezfw_backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/configs/xray": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves a paginated list of Xray configurations.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ConfigsXray"
                ],
                "summary": "List Xray Configs",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "example": 10,
                        "description": "Number of items to return per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "example": 0,
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of Xray configurations",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/github_com_tools4net_ezfw_backend_internal_models.XrayConfig"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Failed to list configurations",
                        "schema": {
                            "$ref": "#/definitions/github_com_tools4net_ezfw_backend_internal_models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Adds a new Xray configuration to the store. Name must be unique.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ConfigsXray"
                ],
                "summary": "Create Xray Config",
                "parameters": [
                    {
                        "description": "Xray Configuration Object",
                        "name": "xrayConfig",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_tools4net_ezfw_backend_internal_models.XrayConfig"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created Xray configuration",
                        "schema": {
                            "$ref": "#/definitions/github_com_tools4net_ezfw_backend_internal_models.XrayConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid input or missing name",
                        "schema": {
                            "$ref": "#/definitions/github_com_tools4net_ezfw_backend_internal_models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict - Configuration name already exists",
                        "schema": {
                            "$ref": "#/definitions/github_com_tools4net_ezfw_backend_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Failed to create configuration",
                        "schema": {
                            "$ref": "#/definitions/github_com_tools4net_ezfw_backend_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_tools4net_ezfw_backend_internal_models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Detailed error message"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key for authentication. Prepend with \"Bearer \" if it's a bearer token, otherwise just the key. For this API, just the key is expected.",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}